requirements.txt
urllib3==1.25.3
mysql.connector==2.2.9
postmarker==0.13.0